# CustomerServiceSwagger.yaml
...
swagger: "2.0"
info:
  description: This is a simple API for Customer Endpoints
  version: 1.0.1
  title: AWS Customer Endpoints
host: virtserver.swaggerhub.com
basePath: /nurikkk99/AWS_Shopping_Project_Customer/1.0.0
...
schemes:
  - https
  - http
paths:
  /api/products:
    post:
      tags:
        - product-controller
      summary: Filtering products by parameters
      description: Returns filtered products
      operationId: filterProducts
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Search products by list of fields and term, filter products by type,manufacturer, ascendind or descending sort by price or date
          required: false
          schema:
            $ref: '#/definitions/SearchAndFilterModel'
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
  /api/products/{productId}:
    get:
      tags:
        - product-controller
      summary: Find product by Id
      description: Returns a single product
      operationId: getProductById
      produces:
        - application/json
      parameters:
        - name: productId
          in: path
          description: ID of a product to return
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
...
  /api/users/{userId}/baskets/{basketId}/products/{productId}:
    put:
      tags:
        - basket-controller
      summary: Add product to basket
      operationId: putBasket
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: User Id
          required: true
          type: integer
          format: int64
        - name: basketId
          in: path
          description: Basket Id
          required: true
          type: string
        - name: productId
          in: path
          description: Product Id
          required: true
          type: string
      responses:
        "201":
          description: Operation succesful
    delete:
      tags:
        - basket-controller
      summary: Delete product from basket
      operationId: deleteFromBasket
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: User Id
          required: true
          type: integer
          format: int64
        - name: basketId
          in: path
          description: Basket Id
          required: true
          type: string
        - name: productId
          in: path
          description: Product Id
          required: true
          type: string
      responses:
        "201":
          description: operation succesful
...
definitions:
  SearchAndFilterModel:
    type: object
    properties:
      searchFields:
        type: array
        items:
          type: string
        example: [name, definition]
      searchTerm:
        type: string
        example: red sneakers
      type:
        type: string
        example: Sneakers
      manufacturer:
        type: string
        example: Nike
      sortType:
        type: string
        example: price, date
      sortOrder:
        type: string
        example: ASC, DESC
  Product:
    type: object
    required:
      - id
      - manufacturer
      - name
      - price
      - releaseDate
      - type
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: AirForce
      type:
        type: string
        example: Sneakers
      price:
        type: integer
        format: int64
        example: 3000
      manufacturer:
        type: string
        example: Nike
      releaseDate:
        type: string
        format: date-time
        example: 2016-08-29T09:12:33.001+0000
...